+++ tests/resources.py
@@ -10,6 +10,8 @@ NEW_UV_PYPROJECT = _RESOURCES / "new-uv-project/pyproject.toml"
 OLD_POETRY_LOCKFILE = _RESOURCES / "old-poetry-project/poetry.lock"
 NEW_POETRY_LOCKFILE = _RESOURCES / "new-poetry-project/poetry.lock"
 NEW_POETRY_PYPROJECT = _RESOURCES / "new-poetry-project/pyproject.toml"
+SOURCES_POETRY_LOCKFILE = _RESOURCES / "sources-poetry/poetry.lock"
+SOURCES_UV_LOCKFILE = _RESOURCES / "sources-uv/uv.lock"
 
 README = _RESOURCES / "../README.md"
 EMPTY = pathlib.Path("/dev/null")